Clean up additional warnings.
The VM now builds cleanly with -Wall enabled. However I have explicitly
disabled 7 warning types that will require significant code refactoring
and/or increased understanding of the internals of the VM.

The VM seems to run and loads boot.rbl with only one warning.

+++ rosette/src/Number.cc
@@ -314,7 +314,7 @@ DEF("fx-lsl", fxLsl, 2, 2) {
 DEF("fx-lsr", fxLsr, 2, 2) {
     CHECK_FIXNUM(0, m);
     CHECK_FIXNUM(1, n);
-    unsigned long k = m;
+    u_int32_t k = m;
     return FIXNUM(k >> n);
 }
